chore: :art: refactor on Pagination module
