+++ src/main/java/com/college/hotlittlepigs/box/BoxService.java
@@ -4,8 +4,8 @@ import com.college.hotlittlepigs.box.dto.BoxUpdateDTO;
 import com.college.hotlittlepigs.box.exception.BoxNotFoundException;
 import com.college.hotlittlepigs.box.exception.ParametersNotFoundException;
 import com.college.hotlittlepigs.gestation.Gestation;
-import com.college.hotlittlepigs.model.PageModel;
-import com.college.hotlittlepigs.model.PageRequestModel;
+import com.college.hotlittlepigs.pagination.PageModel;
+import com.college.hotlittlepigs.pagination.PageRequestModel;
 import com.college.hotlittlepigs.parameters.Parameters;
 import com.college.hotlittlepigs.parameters.ParametersService;
 import lombok.AllArgsConstructor;
