+++ src/main/java/com/college/hotlittlepigs/log_controllers/LogControllersService.java
@@ -1,8 +1,8 @@
 package com.college.hotlittlepigs.log_controllers;
 
 import com.college.hotlittlepigs.controllers.Controller;
-import com.college.hotlittlepigs.model.PageModel;
-import com.college.hotlittlepigs.model.PageRequestModel;
+import com.college.hotlittlepigs.pagination.PageModel;
+import com.college.hotlittlepigs.pagination.PageRequestModel;
 import lombok.AllArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
