+++ src/main/java/com/college/hotlittlepigs/matrix/MatrixService.java
@@ -1,8 +1,8 @@
 package com.college.hotlittlepigs.matrix;
 
 import com.college.hotlittlepigs.matrix.exception.MatrixNotFoundException;
-import com.college.hotlittlepigs.model.PageModel;
-import com.college.hotlittlepigs.model.PageRequestModel;
+import com.college.hotlittlepigs.pagination.PageModel;
+import com.college.hotlittlepigs.pagination.PageRequestModel;
 import lombok.AllArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
