+++ src/main/java/com/college/hotlittlepigs/meansurement/MeasurementResource.java
@@ -1,8 +1,8 @@
 package com.college.hotlittlepigs.meansurement;
 
 import com.college.hotlittlepigs.meansurement.dto.MeasurementListSaveDTO;
-import com.college.hotlittlepigs.model.PageModel;
-import com.college.hotlittlepigs.model.PageRequestModel;
+import com.college.hotlittlepigs.pagination.PageModel;
+import com.college.hotlittlepigs.pagination.PageRequestModel;
 import lombok.AllArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
