+++ src/main/java/com/college/hotlittlepigs/sensors/SensorsResource.java
@@ -4,8 +4,8 @@ import java.util.Map;
 
 import javax.validation.Valid;
 
-import com.college.hotlittlepigs.model.PageModel;
-import com.college.hotlittlepigs.model.PageRequestModel;
+import com.college.hotlittlepigs.pagination.PageModel;
+import com.college.hotlittlepigs.pagination.PageRequestModel;
 import com.college.hotlittlepigs.sensors.dto.SensorsSaveDTO;
 
 import org.springframework.http.HttpStatus;
