+++ src/main/java/com/college/hotlittlepigs/sensors/SensorsService.java
@@ -3,8 +3,8 @@ package com.college.hotlittlepigs.sensors;
 import java.util.Optional;
 
 import com.college.hotlittlepigs.exception.common.NotFoundException;
-import com.college.hotlittlepigs.model.PageModel;
-import com.college.hotlittlepigs.model.PageRequestModel;
+import com.college.hotlittlepigs.pagination.PageModel;
+import com.college.hotlittlepigs.pagination.PageRequestModel;
 
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
