+++ src/test/java/com/college/hotlittlepigs/Repository/GestationRepositoryTests.java
@@ -11,7 +11,7 @@ import com.college.hotlittlepigs.gestation.Gestation;
 import com.college.hotlittlepigs.gestation.GestationRepository;
 import com.college.hotlittlepigs.matrix.Matrix;
 import com.college.hotlittlepigs.matrix.MatrixRepository;
-import com.college.hotlittlepigs.model.PageRequestModel;
+import com.college.hotlittlepigs.pagination.PageRequestModel;
 
 import org.junit.FixMethodOrder;
 import org.junit.Test;
